+++ Libraries/UnitsOfMeasurement/Temperature/Kelvin.cs
@@ -6,7 +6,7 @@
         {
             public class Kelvin : Temperature
             {
-                public Kelvin(decimal value) : base(value, "K") { Value = value; }
+                public Kelvin(double value) : base(value, "K") { Value = value; }
 
                 public static Kelvin operator +(Kelvin firstMeasurement, Kelvin secondMeasurement)
                 {
@@ -34,9 +34,9 @@
             public static Kelvin Kelvins(this int input) => new Kelvin(input);
             public static Kelvin Kelvins(this long input) => new Kelvin(input);
 
-            public static Kelvin Kelvins(this float input) => new Kelvin((decimal)input);
-            public static Kelvin Kelvins(this double input) => new Kelvin((decimal)input);
-            public static Kelvin Kelvins(this decimal input) => new Kelvin(input);
+            public static Kelvin Kelvins(this float input) => new Kelvin((double)input);
+            public static Kelvin Kelvins(this double input) => new Kelvin((double)input);
+            public static Kelvin Kelvins(this double input) => new Kelvin(input);
         }
     }
 }
